requirements.txt
setuptools~=65.5.1
python-dotenv~=1.0.0